% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SpaGene.R
\name{FindPattern}
\alias{FindPattern}
\title{Find spatial patterns}
\usage{
FindPattern(spagene_res, cutoff = 0.01, nPattern = 8)
}
\arguments{
\item{spagene_res}{result from SpaGene, a list containing normexp and spagene_res}

\item{cutoff}{the adjp cutoff to select spatially variable genes (default: 0.01)}

\item{nPattern}{the number of patterns (default:8)}
}
\value{
a list containing the pattern (pattern), gene similarity with the pattern (genepattern), and the pattern weight (patternw)
}
\description{
Find spatial patterns
}
