% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SpaGene.R
\name{FindPattern_Multi}
\alias{FindPattern_Multi}
\title{Find spatial patterns across multiple samples}
\usage{
FindPattern_Multi(spagene_list, cutoff = 0.01, nPattern = 12)
}
\arguments{
\item{spagene_list}{a list of results from SpaGene, each result containing normexp and spagene_res}

\item{cutoff}{the adjp cutoff to select spatially variable genes (default: 0.01)}

\item{nPattern}{the number of patterns (default:12)}
}
\value{
a list containing the pattern (pattern), gene similarity with the pattern (genepattern), and the pattern weight (patternw)
}
\description{
Find spatial patterns across multiple samples
}
