% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SpaGene.R
\name{LRactivity}
\alias{LRactivity}
\title{calculate the activity for each LR pair}
\usage{
LRactivity(expr, location, normalize = T, knn = 8, LRpair = LRpair)
}
\arguments{
\item{expr}{gene expression matrix, the row is the gene and the column is the spot/cell}

\item{location}{location matrix, the row number of location should match the column number of expr}

\item{normalize}{whether to normalize the data (default: TRUE)}

\item{knn}{the number of nearest neighbours to search (default: 8)}

\item{LRpair}{the ligand-receptor pair for plot}
}
\value{
a data matrix with LR activity in each location, row is LR pair, column is location.
}
\description{
calcuate the LR activity
}
