% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SpaGene.R
\name{PlotPattern}
\alias{PlotPattern}
\title{Plot patterns}
\usage{
PlotPattern(pattern, location, max.cutoff = 0.9, pt.size = 2, alpha.min = 0.1)
}
\arguments{
\item{pattern}{pattern result from FindPattern}

\item{location}{location matrix}

\item{max.cutoff}{the maximum value cutoff (default:0.9)}

\item{pt.size}{the point size (default:2)}

\item{alpha.min}{the alpha value of the minimum value (default:0.1)}
}
\value{
a list of ggplot
}
\description{
plot patterns from spatially variable genes
}
