% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SpaGene.R
\name{PlotPattern_Multi}
\alias{PlotPattern_Multi}
\title{Plot patterns across multiple samples}
\usage{
PlotPattern_Multi(
  pattern,
  locationlist,
  patternid = 1,
  max.cutoff = 0.9,
  pt.size = 2,
  alpha.min = 0.1
)
}
\arguments{
\item{pattern}{pattern result from FindPattern_Multi}

\item{max.cutoff}{the maximum value cutoff (default:0.9)}

\item{pt.size}{the point size (default:2)}

\item{alpha.min}{the alpha value of the minimum value (default:0.1)}

\item{location}{a list of location matrix}
}
\value{
a list of ggplot
}
\description{
plot patterns from spatially variable genes from multiple samples
}
